+++ MyPatient.Models/Patient.cs
@@ -34,7 +34,7 @@ namespace MyPatient.Models
         // **********MA***********
         [Required(ErrorMessage = "Debe seleccionar un MA.")]
         [ForeignKey("MA")]
-        public int MAId { get; set; }
+        public long MAId { get; set; }
 
         public Doctor? MA { get; set; }
     }
